+++ src/main/kotlin/lotto/domain/RankResult.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class RankResult(
     val totalCount: Int,
