+++ src/main/kotlin/lotto/util/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 interface NumberGenerator {
     fun generate(): Set<Int>
