+++ src/main/kotlin/lotto/util/NumberSplitter.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 object NumberSplitter {
     private const val SEPARATOR = ","
