+++ src/main/kotlin/lotto/view/ResultView.kt
@@ -1,4 +1,7 @@
-package lotto
+package lotto.view
+
+import lotto.domain.Lotto
+import lotto.domain.WinningResult
 
 object ResultView {
     fun printCreatedLottos(lottos: List<Lotto>) {
