+++ src/test/kotlin/lotto/FixedNumberGenerator.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.util.NumberGenerator
+
 class FixedNumberGenerator : NumberGenerator {
     override fun generate(): Set<Int> {
         return setOf(1, 2, 3, 4, 5, 6)
